backend/requirements/base.txt
boto3==1.9.130

celery==4.3.0

channels==2.2.0
channels-redis==2.4.0

daphne==2.3.0

graphene-django==2.8.1
django-graphql-jwt==0.3.0

Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework_simplejwt==4.3.0
django-shortcuts==1.6
django-debug-toolbar==1.11


# Python Social Auth

social-auth-app-django==3.1.0
social-auth-core==3.2.0

# Django page caching is not working with this version of Redis
# but this version is required by Celery
django-redis==4.10.0

django-storages==1.7.1
django-constance==2.4.0 # for managing settings in admin

pandas==0.24.2
pandas_flavor==0.1.2
pyjanitor==0.16.6

xlrd==1.2.0 # required by pandas for Excel support
XlsxWriter==1.1.6 # used by the excel recon tool
django-postgres-copy==2.4.0
django-safedelete==0.5.1

gunicorn==19.9.0

psycopg2-binary==2.8.2

redis==3.2.1

requests==2.21.0